More fixes

+++ test_functions.sh
@@ -148,8 +148,8 @@ function create_and_cd_test_dir {
     local out_file
     while IFS= read -r -d $'\0' file; do
         out_file="$(basename "${file}")"
-        if [[ "${out_file}" == "target" ]]; then
-            continue # Don't copy target directories
+        if [[ "${out_file}" == "target" || "${out_file}" == "Cargo.lock" ]]; then
+            continue # Not sharing these files is exactly why we're creating a separate test dir
         fi
         out_file="${target_dir}/${out_file}"
         rm -f "${out_file}" # Remove any existing files
@@ -269,7 +269,7 @@ function _internal_run_error_message_tests {
 
     [[ ${error} -eq 0 ]] || return 1
 
-    mkdir -p "${COVERAGE_DIR:?}"/{error_messages_stable,error_messages_nightly}
+    mkdir -p target/cov/{err_stable,err_nightly}
 
     # Run the tests
     if [[ ${overwrite} -eq 1 ]]; then
@@ -282,18 +282,18 @@ function _internal_run_error_message_tests {
         assert_no_change "${fail_dir}" || return 1
 
         # Run stable tests
-        try_silent cargo +stable llvm-cov test error_message_tests --workspace --lcov --output-path "${COVERAGE_DIR}/error_messages_stable/lcov.info" -- --ignored || exit 1
+        try_silent cargo +stable llvm-cov test error_message_tests --workspace --no-clean --lcov --output-path target/cov/err_stable/lcov.info -- --ignored || exit 1
 
         assert_no_change "${fail_dir}" || return 1
 
         # Run nightly tests
-        try_silent cargo +nightly llvm-cov test error_message_tests --workspace --lcov --output-path "${COVERAGE_DIR}/error_messages_nightly/lcov.info" -- --ignored || exit 1
+        try_silent cargo +nightly llvm-cov test error_message_tests --workspace --no-clean --lcov --output-path target/cov/err_nightly/lcov.info -- --ignored || exit 1
 
         assert_no_change "${fail_dir}" "nightly" || return 1
     else
         unset TRYBUILD # Remove TRYBUILD flag if it was set
-        try_silent cargo +stable llvm-cov test error_message_tests --workspace --lcov --output-path "${COVERAGE_DIR}/error_messages_stable/lcov.info" -- --ignored || exit 1
-        try_silent cargo +nightly llvm-cov test error_message_tests --workspace --lcov --output-path "${COVERAGE_DIR}/error_messages_nightly/lcov.info" -- --ignored || exit 1
+        try_silent cargo +stable llvm-cov test error_message_tests --workspace --no-clean --lcov --output-path target/cov/err_stable/lcov.info -- --ignored || exit 1
+        try_silent cargo +nightly llvm-cov test error_message_tests --workspace --no-clean --lcov --output-path target/cov/err_nightly/lcov.info -- --ignored || exit 1
     fi
 
     # Check that the stable and nightly distinction is actually used
